Add support for dash client-side callbacks (#149) (#165)
This should close issue #149. This is a purely additive change.

Adding support for client-side callbacks is just a matter of registering a javascript function name. The client-side callback may be registered by calling the `clientside_callback` function on a `DjangoDash` object in the same exact way as it would have been done with a plain Dash app (see: https://community.plot.ly/t/dash-0-41-0-released/22131).